code reformatting

+++ Compression.BSA.Test/Program.cs
@@ -1,33 +1,28 @@
-﻿using ICSharpCode.SharpZipLib;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Compression.BSA.Test
 {
-    class Program
+    internal class Program
     {
-        const string TestDir = @"D:\MO2 Instances\";
-        const string TempDir = "c:\\tmp\\out";
-        static void Main(string[] args)
+        private const string TestDir = @"D:\MO2 Instances\";
+        private const string TempDir = "c:\\tmp\\out";
+
+        private static void Main(string[] args)
         {
             foreach (var bsa in Directory.EnumerateFiles(TestDir, "*.bsa", SearchOption.AllDirectories).Skip(0))
             {
                 Console.WriteLine($"From {bsa}");
                 Console.WriteLine("Cleaning Output Dir");
-                if (Directory.Exists(TempDir))
-                {
-                    Directory.Delete(TempDir, true);
-                }
+                if (Directory.Exists(TempDir)) Directory.Delete(TempDir, true);
                 Directory.CreateDirectory(TempDir);
 
                 Console.WriteLine($"Reading {bsa}");
                 using (var a = new BSAReader(bsa))
                 {
-
                     Parallel.ForEach(a.Files, file =>
                     {
                         var abs_name = Path.Combine(TempDir, file.Path);
@@ -36,8 +31,11 @@ namespace Compression.BSA.Test
                             Directory.CreateDirectory(Path.GetDirectoryName(abs_name));
 
                         using (var fs = File.OpenWrite(abs_name))
+                        {
                             file.CopyDataTo(fs);
-                        Equal((long)file.Size, new FileInfo(abs_name).Length);
+                        }
+
+                        Equal(file.Size, new FileInfo(abs_name).Length);
                     });
 
 
@@ -56,9 +54,8 @@ namespace Compression.BSA.Test
                             {
                                 var entry = w.AddFile(file.Path, str, file.FlipCompression);
                             }
-
                         });
-                        
+
                         w.Build("c:\\tmp\\tmp.bsa");
 
                         // Sanity Checks
@@ -70,33 +67,31 @@ namespace Compression.BSA.Test
                             Console.WriteLine($"{pair.ai.Path},  {pair.ai.Hash},  {pair.bi.Path}, {pair.bi.Hash}");
                         }*/
 
-                        foreach (var pair in Enumerable.Zip(a.Files, w.Files, (ai, bi) => (ai, bi)))
+                        foreach (var pair in a.Files.Zip(w.Files, (ai, bi) => (ai, bi)))
                         {
                             Equal(pair.ai.Path, pair.bi.Path);
                             Equal(pair.ai.Hash, pair.bi.Hash);
                         }
-
                     }
 
                     Console.WriteLine($"Verifying {bsa}");
                     using (var b = new BSAReader("c:\\tmp\\tmp.bsa"))
                     {
                         Console.WriteLine($"Performing A/B tests on {bsa}");
-                        Equal((uint)a.ArchiveFlags, (uint)b.ArchiveFlags);
-                        Equal((uint)a.FileFlags, (uint)b.FileFlags);
+                        Equal((uint) a.ArchiveFlags, (uint) b.ArchiveFlags);
+                        Equal((uint) a.FileFlags, (uint) b.FileFlags);
 
                         // Check same number of files
                         Equal(a.Files.Count(), b.Files.Count());
-                        int idx = 0;
-                        foreach (var pair in Enumerable.Zip(a.Files, b.Files, (ai, bi) => (ai, bi)))
+                        var idx = 0;
+                        foreach (var pair in a.Files.Zip(b.Files, (ai, bi) => (ai, bi)))
                         {
-                            idx ++;
+                            idx++;
                             //Console.WriteLine($"   - {pair.ai.Path}");
                             Equal(pair.ai.Path, pair.bi.Path);
                             Equal(pair.ai.Compressed, pair.bi.Compressed);
                             Equal(pair.ai.Size, pair.bi.Size);
                             Equal(pair.ai.GetData(), pair.bi.GetData());
-
                         }
                     }
                 }
@@ -109,7 +104,6 @@ namespace Compression.BSA.Test
 
             foreach (var itm in a)
                 Equal(b.Contains(itm));
-
         }
 
         private static void Equal(bool v)
@@ -161,16 +155,11 @@ namespace Compression.BSA.Test
 
         public static void Equal(byte[] a, byte[] b)
         {
-            if (a.Length != b.Length)
-            {
-                throw new InvalidDataException($"Byte array sizes are not equal");
-            }
+            if (a.Length != b.Length) throw new InvalidDataException("Byte array sizes are not equal");
 
-            for (var idx = 0; idx < a.Length; idx ++)
-            {
+            for (var idx = 0; idx < a.Length; idx++)
                 if (a[idx] != b[idx])
                     throw new InvalidDataException($"Byte array contents not equal at {idx}");
-            }
         }
     }
 }
